+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - #186 - Add CheckStyle
 - #156 - Add exception handler to starter
 - #193 - Checking and add uniqueness_restrictions in the database
+- #198 - Add sorting of payments by ID
 
 ## [1.4.0] - 2.04.2024
 - #167 - Testing the conversion of Product review to productDTO rating